-changement des constantes pour mettre trois constantes TAILLE50, TAILLE32 et TAILLE16. -changement de la méthode déposer pour déposer les objets au pied du personnage

+++ src/main/java/universite_paris8/iut/yponnou/zelda/modele/Armes/Arme.java
@@ -8,7 +8,7 @@ public class Arme extends Objet {
     private int ptsDegats;
     private int portee;
 
-    public Arme(String nom, int x, int y, Environnement env, int ptsDegats, int portee) {
+    public Arme(String nom, double x, double y, Environnement env, int ptsDegats, int portee) {
         super(x,y,env);
         this.nom = nom;
         this.ptsDegats = ptsDegats;
