+++ src/main/java/universite_paris8/iut/yponnou/zelda/modele/Armes/ArmeDistance.java
@@ -5,7 +5,7 @@ import universite_paris8.iut.yponnou.zelda.modele.Environnement;
 public class ArmeDistance extends Arme{
 
 
-    public ArmeDistance(String nom, int x, int y, Environnement env, int ptsDegats, int portee) {
+    public ArmeDistance(String nom, double x, double y, Environnement env, int ptsDegats, int portee) {
         super(nom, x, y, env, ptsDegats, portee);
     }
 }
