+++ src/main/java/universite_paris8/iut/yponnou/zelda/modele/Armes/Epee.java
@@ -4,9 +4,7 @@ import universite_paris8.iut.yponnou.zelda.modele.Environnement;
 
 public class Epee extends ArmeMelee{
 
-
-    public Epee(int x, int y, Environnement env) {
+    public Epee(double x, double y, Environnement env) {
         super("Epee", x, y, env,15, 5);
     }
-
 }
