+++ src/main/java/universite_paris8/iut/yponnou/zelda/modele/Objets/Aliments/Pomme.java
@@ -4,7 +4,7 @@ import universite_paris8.iut.yponnou.zelda.modele.Environnement;
 
 public class Pomme extends Nourriture {
 
-    public Pomme(int x, int y, Environnement environnement) {
+    public Pomme(double x, double y, Environnement environnement) {
         super(x, y, environnement, 5);
         setNom("Pomme");
     }
